// packages/blog-starter-kit/themes/enterprise/generated/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type DarkModePreferences = {
	__typename?: 'DarkModePreferences';
	/** A flag indicating if the dark mode is enabled for the publication. */
	enabled?: Maybe<Scalars['Boolean']['output']>;
	/** The custom dark mode logo of the publication. */
	logo?: Maybe<Scalars['String']['output']>;
};
// ...
export type OpenGraphMetaData = {
	__typename?: 'OpenGraphMetaData';
	/** The image used in og:image tag for SEO purposes. */
	image?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageInfo = {
	__typename?: 'PageInfo';
	/**
	 * The cursor of the last item in the current page.
	 * Use it as the after input to query the next page.
	 */
	endCursor?: Maybe<Scalars['String']['output']>;
	/** Indicates if there are more pages. */
	hasNextPage?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type PostCoverImage = {
	__typename?: 'PostCoverImage';
	/** Provides attribution information for the cover image, if available. */
	attribution?: Maybe<Scalars['String']['output']>;
	/** True if the image attribution should be hidden. */
	isAttributionHidden: Scalars['Boolean']['output'];
	/** Indicates whether the cover image is in portrait orientation. */
	isPortrait: Scalars['Boolean']['output'];
	/** The name of the photographer who captured the cover image. */
	photographer?: Maybe<Scalars['String']['output']>;
	/** The URL of the cover image. */
	url: Scalars['String']['output'];
};

/** Contains a post and a cursor for pagination. */
export type PostEdge = Edge & {
	__typename?: 'PostEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String']['output'];
	/** The node holding the Post information */
	node: Post;
};
// ...
export type Preferences = {
	__typename?: 'Preferences';
	/** The publication's darkmode preferences. Can be used to load blog in dark mode by default and add a custom dark mode logo. */
	darkMode?: Maybe<DarkModePreferences>;
	/** A flag indicating if the hashnode's footer branding is disabled for the publication. */
	disableFooterBranding?: Maybe<Scalars['Boolean']['output']>;
	/** An object containing pages enabled for the publication. */
	enabledPages?: Maybe<PagesPreferences>;
	/** A flag indicating if subscription popup needs to be shown to be shown for the publication */
	isSubscriptionModalDisabled?: Maybe<Scalars['Boolean']['output']>;
	/** The selected publication's layout, can be stacked, grid or magazine. */
	layout?: Maybe<PublicationLayout>;
	/** The publication's logo url. */
	logo?: Maybe<Scalars['String']['output']>;
	/** The items in the publication's navigation bar. */
	navbarItems: Array<PublicationNavbarItem>;
};
// ...
export type PublicationNavbarItem = {
	__typename?: 'PublicationNavbarItem';
	/** The unique identifier of the navbar item. */
	id: Scalars['ID']['output'];
	/** The label of the navbar item. */
	label?: Maybe<Scalars['String']['output']>;
	/** The static page added to the navbar item. */
	page?: Maybe<StaticPage>;
	/**
	 * The order of the navbar item.
	 * @deprecated Navbar items are already returned in the correct order. Priority value is not needed and might be 0 in most cases.
	 */
	priority?: Maybe<Scalars['Int']['output']>;
	/** The series added to the navbar item. */
	series?: Maybe<Series>;
	/** The type of the navbar item, can be series, link or page. */
	type: PublicationNavigationType;
	/** The URL of the navbar item. */
	url?: Maybe<Scalars['String']['output']>;
};
// ...
export type PublicationPostConnection = Connection & {
	__typename?: 'PublicationPostConnection';
	/** A list of edges containing Post information */
	edges: Array<PostEdge>;
	/** Information for pagination in Post connection. */
	pageInfo: PageInfo;
	/** The total number of documents in the connection. */
	totalDocuments: Scalars['Int']['output'];
};
// ...
export type Query = {
	__typename?: 'Query';
	checkCustomDomainAvailability: CheckCustomDomainAvailabilityResult;
	checkSubdomainAvailability: CheckSubdomainAvailabilityResult;
	documentationProject?: Maybe<DocumentationProject>;
	/**
	 * Returns a draft by ID.
	 * Draft is a post that is not published yet.
	 */
	draft?: Maybe<Draft>;
	/**
	 * Returns a paginated list of posts based on the provided filter.
	 * Used in Hashnode home feed.
	 */
	feed: FeedPostConnection;
	/** Returns the current authenticated user. Only available to the authenticated user. */
	me: MyUser;
	/** Returns post by ID. Can be used to render post page on blog. */
	post?: Maybe<Post>;
	/**
	 * Returns the publication with the given ID or host.
	 * User can pass anyone of them.
	 */
	publication?: Maybe<Publication>;
	/** Get a scheduled post by ID. */
	scheduledPost?: Maybe<ScheduledPost>;
	/** Returns a paginated list of posts based on search query for a particular publication id. */
	searchPostsOfPublication: SearchPostConnection;
	/** Returns tag details by its slug. */
	tag?: Maybe<Tag>;
	/** Returns users who have most actively participated in discussions by commenting in the last 7 days. */
	topCommenters: CommenterUserConnection;
	/** Returns the user with the username. */
	user?: Maybe<User>;
};
// ...
export type User = IUser &
	Node & {
		__typename?: 'User';
		/**
		 * Whether or not the user is an ambassador.
		 * @deprecated Ambassadors program no longer active. Will be removed after 02/01/2024
		 */
		ambassador: Scalars['Boolean']['output'];
		/** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
		availableFor?: Maybe<Scalars['String']['output']>;
		/** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
		badges: Array<Badge>;
		/** The bio of the user. Visible in about me section of the user's profile. */
		bio?: Maybe<Content>;
		/**
		 * The bio of the user. Visible in about me section of the user's profile.
		 * @deprecated Will be removed on 26/10/2023. Use bio instead of bioV2
		 */
		bioV2?: Maybe<Content>;
		/** The date the user joined Hashnode. */
		dateJoined?: Maybe<Scalars['DateTime']['output']>;
		/** Whether or not the user is deactivated. */
		deactivated: Scalars['Boolean']['output'];
		/** The users who are following this user */
		followers: UserConnection;
		/** The number of users that follow the requested user. Visible in the user's profile. */
		followersCount: Scalars['Int']['output'];
		/**
		 * Whether or not the authenticated user follows this user.
		 * Returns false if the authenticated user this user.
		 */
		following: Scalars['Boolean']['output'];
		/** The number of users that this user is following. Visible in the user's profile. */
		followingsCount: Scalars['Int']['output'];
		/** The users which this user is following */
		follows: UserConnection;
		/**
		 * Whether or not this user follows the authenticated user.
		 * Returns false if the authenticated user this user.
		 */
		followsBack: Scalars['Boolean']['output'];
		/** The ID of the user. It can be used to identify the user. */
		id: Scalars['ID']['output'];
		/** The location of the user. */
		location?: Maybe<Scalars['String']['output']>;
		/** The name of the user. */
		name: Scalars['String']['output'];
		/** Returns the list of posts the user has published. */
		posts: UserPostConnection;
		/** The URL to the profile picture of the user. */
		profilePicture?: Maybe<Scalars['String']['output']>;
		/** Publications associated with the user. Includes personal and team publications. */
		publications: UserPublicationsConnection;
		/** The social media links of the user. Shown on the user's profile. */
		socialMediaLinks?: Maybe<SocialMediaLinks>;
		/** The tagline of the user. Shown on the user's profile below the name. */
		tagline?: Maybe<Scalars['String']['output']>;
		/** Returns a list of tags that the user follows. */
		tagsFollowing: Array<Tag>;
		/** Returns list of tags from user's expertise. Shown on the user's profile. */
		techStack: UserTagsConnection;
		/** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
		username: Scalars['String']['output'];
	};
// ...
export type PublicationFragment = {
	__typename?: 'Publication';
	id: string;
	title: string;
	displayTitle?: string | null;
	url: string;
	metaTags?: string | null;
	favicon?: string | null;
	isTeam: boolean;
	followersCount?: number | null;
	descriptionSEO?: string | null;
	author: {
		__typename?: 'User';
		name: string;
		username: string;
		profilePicture?: string | null;
		followersCount: number;
	};
	ogMetaData: { __typename?: 'OpenGraphMetaData'; image?: string | null };
	preferences: {
		__typename?: 'Preferences';
		logo?: string | null;
		darkMode?: { __typename?: 'DarkModePreferences'; logo?: string | null } | null;
		navbarItems: Array<{
			__typename?: 'PublicationNavbarItem';
			id: string;
			type: PublicationNavigationType;
			label?: string | null;
			url?: string | null;
		}>;
	};
	links?: {
		__typename?: 'PublicationLinks';
		twitter?: string | null;
		github?: string | null;
		linkedin?: string | null;
		hashnode?: string | null;
	} | null;
	integrations?: {
		__typename?: 'PublicationIntegrations';
		umamiWebsiteUUID?: string | null;
		gaTrackingID?: string | null;
		fbPixelID?: string | null;
		hotjarSiteID?: string | null;
		matomoURL?: string | null;
		matomoSiteID?: string | null;
		fathomSiteID?: string | null;
		gTagManagerID?: string | null;
		fathomCustomDomain?: string | null;
		fathomCustomDomainEnabled?: boolean | null;
		plausibleAnalyticsEnabled?: boolean | null;
		koalaPublicKey?: string | null;
		msClarityID?: string | null;
	} | null;
};
// ...
export type PostsByPublicationQueryVariables = Exact<{
	host: Scalars['String']['input'];
	first: Scalars['Int']['input'];
	after?: InputMaybe<Scalars['String']['input']>;
}>;

export type PostsByPublicationQuery = {
	__typename?: 'Query';
	publication?: {
		__typename?: 'Publication';
		id: string;
		title: string;
		displayTitle?: string | null;
		url: string;
		metaTags?: string | null;
		favicon?: string | null;
		isTeam: boolean;
		followersCount?: number | null;
		descriptionSEO?: string | null;
		posts: {
			__typename?: 'PublicationPostConnection';
			totalDocuments: number;
			edges: Array<{
				__typename?: 'PostEdge';
				node: {
					__typename?: 'Post';
					id: string;
					title: string;
					url: string;
					publishedAt: string;
					slug: string;
					brief: string;
					author: { __typename?: 'User'; name: string; profilePicture?: string | null };
					coverImage?: { __typename?: 'PostCoverImage'; url: string } | null;
				};
			}>;
			pageInfo: {
				__typename?: 'PageInfo';
				endCursor?: string | null;
				hasNextPage?: boolean | null;
			};
		};
		author: {
			__typename?: 'User';
			name: string;
			username: string;
			profilePicture?: string | null;
			followersCount: number;
		};
		ogMetaData: { __typename?: 'OpenGraphMetaData'; image?: string | null };
		preferences: {
			__typename?: 'Preferences';
			logo?: string | null;
			darkMode?: { __typename?: 'DarkModePreferences'; logo?: string | null } | null;
			navbarItems: Array<{
				__typename?: 'PublicationNavbarItem';
				id: string;
				type: PublicationNavigationType;
				label?: string | null;
				url?: string | null;
			}>;
		};
		links?: {
			__typename?: 'PublicationLinks';
			twitter?: string | null;
			github?: string | null;
			linkedin?: string | null;
			hashnode?: string | null;
		} | null;
		integrations?: {
			__typename?: 'PublicationIntegrations';
			umamiWebsiteUUID?: string | null;
			gaTrackingID?: string | null;
			fbPixelID?: string | null;
			hotjarSiteID?: string | null;
			matomoURL?: string | null;
			matomoSiteID?: string | null;
			fathomSiteID?: string | null;
			gTagManagerID?: string | null;
			fathomCustomDomain?: string | null;
			fathomCustomDomainEnabled?: boolean | null;
			plausibleAnalyticsEnabled?: boolean | null;
			koalaPublicKey?: string | null;
			msClarityID?: string | null;
		} | null;
	} | null;
};